move src/shutdown.rs -> src/io/shutdown.rs
