+++ tokio-io/src/io/mod.rs
@@ -15,6 +15,7 @@ mod read;
 mod read_exact;
 mod read_to_end;
 mod read_until;
+mod shutdown;
 
 pub use allow_std::AllowStdIo;
 pub use self::copy::{copy, Copy};
@@ -24,7 +25,7 @@ pub use self::read::{read, Read};
 pub use self::read_exact::{read_exact, ReadExact};
 pub use self::read_to_end::{read_to_end, ReadToEnd};
 pub use self::read_until::{read_until, ReadUntil};
-pub use shutdown::{shutdown, Shutdown};
+pub use self::shutdown::{shutdown, Shutdown};
 pub use split::{ReadHalf, WriteHalf};
 pub use window::Window;
 pub use write_all::{write_all, WriteAll};
